Fix broken WKOF usage
WKOF can't be referenced anymore without using unsafeWindow as far as I can tell, at least in Chrome, so the reference wasn't working.

I've changed it to use unsafeWindow, and also not force an alert. The script should work as it normally did without the framework instead of annoying people to use it just for that one feature.